requirements.txt
PyQt6>=6.6.0
PyYAML>=6.0
Pygments>=2.17.0
Markdown>=3.5.0
